+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_package.py
@@ -16,6 +16,7 @@ import torch
 from lightly_train._data import cache
 from lightly_train._models import log_usage_example
 from lightly_train._models.dinov2_vit.dinov2_vit import DINOv2ViTModelWrapper
+from lightly_train._models.dinov2_vit.dinov2_vit_src import dinov2_helper
 from lightly_train._models.dinov2_vit.dinov2_vit_src.configs import (
     MODELS as VIT_MODELS,
 )
@@ -23,7 +24,6 @@ from lightly_train._models.dinov2_vit.dinov2_vit_src.configs import (
     get_config_path,
     load_and_merge_config,
 )
-from lightly_train._models.dinov2_vit.dinov2_vit_src.dinov2_helper import load_weights
 from lightly_train._models.dinov2_vit.dinov2_vit_src.models import (
     vision_transformer as vits,
 )
@@ -81,6 +81,7 @@ class DINOv2ViTPackage(Package):
         model_name: str,
         num_input_channels: int = 3,
         model_args: dict[str, Any] | None = None,
+        load_weights: bool = True,
     ) -> DinoVisionTransformer:
         """
         Get a DINOv2 ViT model by name. Here the student version is build.
@@ -124,9 +125,9 @@ class DINOv2ViTPackage(Package):
 
         # Load weights if available.
         weights_url = VIT_MODELS[model_name].get("url")
-        if weights_url:
+        if load_weights and weights_url:
             checkpoint_dir = cache.get_model_cache_dir()
-            model = load_weights(
+            model = dinov2_helper.load_weights(
                 model=model,
                 checkpoint_dir=checkpoint_dir,
                 url=weights_url,