+++ src/lightly_train/_task_models/dinov3_eomt_semantic_segmentation/task_model.py
@@ -54,6 +54,7 @@ class DINOv3EoMTSemanticSegmentation(TaskModel):
         backbone_url: str | None = None,
         backbone_weights: PathLike | None = None,
         backbone_args: dict[str, Any] | None = None,
+        load_weights: bool = True,
     ) -> None:
         """
         Args:
@@ -85,12 +86,16 @@ class DINOv3EoMTSemanticSegmentation(TaskModel):
                 The path to the DINOv3 backbone weights. The weights must be exported
                 using LightlyTrain.
             backbone_url:
-                The URL to the DINOv3 backbone weights. This is used to download the
-                weights.
+                Deprecated, weights are now automatically downloaded based on model name.
+                The URL to the DINOv3 backbone weights.
             backbone_args:
                 Additional arguments to pass to the DINOv3 backbone.
+            load_weights:
+                If False, then no pretrained weights are loaded.
         """
-        super().__init__(locals(), ignore_args={"backbone_weights", "backbone_url"})
+        super().__init__(
+            locals(), ignore_args={"backbone_weights", "backbone_url", "load_weights"}
+        )
         parsed_name = self.parse_model_name(model_name=model_name)
         self.model_name = parsed_name["model_name"]
         self.classes = classes
@@ -119,13 +124,8 @@ class DINOv3EoMTSemanticSegmentation(TaskModel):
         backbone_model_args: dict[str, Any] = {
             "in_chans": len(self.image_normalize["mean"]),
         }
-        if backbone_url is not None:
+        if load_weights and backbone_url is not None:
             backbone_model_args["weights"] = backbone_url
-        else:
-            # Set pretrained to false when loading the model for inference. This skips
-            # loading the pretrained weights from Meta as we'll be loading weights with
-            # load_train_state_dict instead.
-            backbone_model_args["pretrained"] = False
         if backbone_args is not None:
             backbone_model_args.update(backbone_args)
 
@@ -133,6 +133,7 @@ class DINOv3EoMTSemanticSegmentation(TaskModel):
         self.backbone = DINOV3_PACKAGE.get_model(
             model_name=parsed_name["backbone_name"],
             model_args=backbone_model_args,
+            load_weights=load_weights,
         )
         assert isinstance(self.backbone, DinoVisionTransformer)
         embed_dim = self.backbone.embed_dim